creating user profile

+++ frontend/components/projects/project_index_item.jsx
@@ -10,13 +10,14 @@ class ProjectIndexItem extends React.Component {
   }
 
   handleDelete(projectId){
-    // debugger
     this.props.deleteProject(projectId);
   }
 
   renderDeleteButton(project){
-    if (this.props.currentUser.id === this.props.project.author_id){
-      return <button className="project-index-item-delete" onClick={()=>this.handleDelete(this.props.project.id)}><i class="far fa-trash-alt"></i></button>
+    if (this.props.currentUser){
+      if (this.props.currentUser.id === this.props.project.author_id){
+        return <button className="project-index-item-delete" onClick={()=>this.handleDelete(this.props.project.id)}><i class="far fa-trash-alt"></i></button>
+      }
     }
   }
